Account for camera rotation when dragging objects.
When panning an individual component, or the whole translationGroup,
remove the camera rotation from the pan vector, so we always pan in
screen XY. This ensures that objects follow the mouse as they're being
dragged.

The pan methods already account for component and scene rotation, but
with the recent addition of the camera rig it's also necessary to
adjust for the camera position.